Counts should be ints

+++ SensitivityModule.h
@@ -109,14 +109,14 @@ typedef struct SensitivityEventStorage{
 
   // Debug information
 
-  double calorimeter_hit_count_; // How many calorimeter hits over threshold?
-  double cluster_count_; // How many clusters with 3 or more hits?
+  int calorimeter_hit_count_; // How many calorimeter hits over threshold?
+  int cluster_count_; // How many clusters with 3 or more hits?
   int track_count_; // How many reconstructed tracks?
   int associated_track_count_; // How many reconstructed tracks with an associated calorimeter?
   int alpha_count_; // How many reconstructed alphas (ie delayed hits)?
   int foil_alpha_count_; //How many reconstructed alphas (ie delayed hits) that we think have a vertex on the foil?
   int delayed_cluster_hit_count_; //How many gieger hits in the alpha track
-  double small_cluster_count_; // How many clusters with 2 hits?
+  int small_cluster_count_; // How many clusters with 2 hits?
   double highest_gamma_energy_; // Highest energy gamma
   double edgemost_vertex_; // Y position of the foil vertex nearest the side of the detector
   double alpha_track_length_; // Length of the alpha track in mm, different metrics for different numbers of delayed hits
